chore: rename store package into path store
